Change field of Point and Lines, change comparingDouble to comparingInt

+++ src/com/company/Point.java
@@ -1,19 +1,19 @@
 package com.company;
 
 public class Point {
-    public final double x;
-    public final double y;
+    public final int x;
+    public final int y;
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
